Add OOP-Features
