+++ data/duke.txt
@@ -1,0 +1,3 @@
+1. T | 1 | b
+2. E | 1 | xd | 11/05/2019 1600-1800
+3. T | 0 | a
